Add publication for marketplace shop packages for admin panel
